Add files via upload
writes palindromic words to a new file named "palindromes.txt". It defines a class called "palindromes" with a main method. The program utilizes BufferedReader to read lines from the input file, FileWriter to write to the output file, and FileReader to read from the input file. The isPalindrome method checks if a given string is a palindrome recursively. If a word in the input file is a palindrome, it is written to the output file.